+++ backend/database.py
@@ -2,24 +2,34 @@ from sqlalchemy import create_engine
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker
 from sqlalchemy.orm import Session
+from dotenv import load_dotenv
+import os
 
+
+load_dotenv()
 
 def construct_db_url(
-    user: str = "postgres",
-    password: str = "1234",
+    type: str = "postgres",
+    user: str = os.environ["DATABASE_USER"],
+    password: str = os.environ["DATABASE_PWD"],
     host: str = "localhost",
     port: str = "5432",
     database: str = "db_dicom_market",
     dialect="postgresql",
-    driver="psycopg2",
 ) -> str:
 
-    return f"{dialect}://{user}:{password}@{host}:{port}/{database}"
+    if (type == "postgres"):
+        return f"{dialect}://{user}:{password}@{host}:{port}/{database}"
+    elif (type == "sqlite"):
+        return "sqlite:///./dcm_test_db.db"
+    else:
+        raise Exception(f"Incorrect type: {type} of database given. The function supports 'sqlite' and 'postgres' types")
+
+SQLALCHEMY_DATABASE_URL = construct_db_url(type = "postgres")
 
 
-def get_connection():
+def get_connection(SQLALCHEMY_DATABASE_URL):
 
-    SQLALCHEMY_DATABASE_URL = construct_db_url()
     try:
         # The Engine is the starting point for any SQLAlchemy application. It’s “home base” for the actual database and its DBAPI
         engine = create_engine(SQLALCHEMY_DATABASE_URL)
@@ -31,7 +41,7 @@ def get_connection():
         )
 
 
-engine = get_connection()
+engine = get_connection(SQLALCHEMY_DATABASE_URL)
 
 
 # each instance of this class will be a Session